Split RSM into package
