+++ src/main/kotlin/org/srcgll/Example.kt
@@ -2,8 +2,7 @@ package org.srcgll
 
 import org.srcgll.grammar.combinator.Grammar
 import org.srcgll.grammar.combinator.regexp.*
-import org.srcgll.grammar.symbol.Terminal
-import org.srcgll.grammar.writeRSMToDOT
+import org.srcgll.rsm.symbol.Terminal
 import org.srcgll.input.Edge
 import org.srcgll.input.ILabel
 import org.srcgll.input.IGraph
