+++ src/main/kotlin/org/srcgll/benchmarks/Benchmarks.kt
@@ -6,14 +6,13 @@ import kotlinx.cli.default
 import kotlinx.cli.required
 import org.srcgll.GLL
 import org.srcgll.RecoveryMode
-import org.srcgll.grammar.readRSMFromTXT
-import org.srcgll.grammar.symbol.Terminal
+import org.srcgll.rsm.readRSMFromTXT
+import org.srcgll.rsm.symbol.Terminal
 import org.srcgll.input.LinearInput
 import org.srcgll.input.LinearInputLabel
 import org.srcgll.lexer.GeneratedLexer
 import org.srcgll.lexer.SymbolCode
 import org.srcgll.lexer.Token
-import org.srcgll.sppf.node.ISPPFNode
 import org.srcgll.sppf.node.SPPFNode
 import org.srcgll.sppf.writeSPPFToDOT
 import java.io.File
