--- src/main/kotlin/org/srcgll/grammar/symbol/Symbol.kt
@@ -1,3 +1,0 @@
-package org.srcgll.grammar.symbol
-
-interface Symbol
