+++ src/main/kotlin/org/srcgll/rsm/RSMNonterminalEdge.kt
@@ -1,6 +1,6 @@
-package org.srcgll.grammar
+package org.srcgll.rsm
 
-import org.srcgll.grammar.symbol.Nonterminal
+import org.srcgll.rsm.symbol.Nonterminal
 
 class RSMNonterminalEdge
 (
