+++ src/main/kotlin/org/srcgll/rsm/RSMTerminalEdge.kt
@@ -1,6 +1,6 @@
-package org.srcgll.grammar
+package org.srcgll.rsm
 
-import org.srcgll.grammar.symbol.Terminal
+import org.srcgll.rsm.symbol.Terminal
 
 class RSMTerminalEdge
 (
