+++ src/main/kotlin/org/srcgll/rsm/RSMWrite.kt
@@ -1,6 +1,6 @@
-package org.srcgll.grammar
+package org.srcgll.rsm
 
-import org.srcgll.grammar.symbol.Nonterminal
+import org.srcgll.rsm.symbol.Nonterminal
 import java.io.File
 
 fun writeRSMToTXT(startState: RSMState, pathToTXT: String) {
