+++ src/main/kotlin/org/srcgll/rsm/symbol/Symbol.kt
@@ -1,0 +1,3 @@
+package org.srcgll.rsm.symbol
+
+interface Symbol
