+++ src/main/kotlin/org/srcgll/sppf/node/PackedSPPFNode.kt
@@ -1,6 +1,6 @@
 package org.srcgll.sppf.node
 
-import org.srcgll.grammar.RSMState
+import org.srcgll.rsm.RSMState
 import java.util.*
 
 open class PackedSPPFNode <VertexType>
