+++ src/test/kotlin/TestRSMStringInputWIthSPPFRecovery.kt
@@ -4,10 +4,11 @@ import org.junit.jupiter.params.provider.MethodSource
 import org.srcgll.GLL
 import org.srcgll.RecoveryMode
 import org.srcgll.sppf.buildStringFromSPPF
-import org.srcgll.grammar.readRSMFromTXT
+import org.srcgll.rsm.readRSMFromTXT
 import org.srcgll.grammar.symbol.*
 import org.srcgll.input.LinearInput
 import org.srcgll.input.LinearInputLabel
+import org.srcgll.rsm.symbol.Terminal
 import org.srcgll.sppf.writeSPPFToDOT
 import kotlin.test.assertNotNull
